+++ cryptofeed/exchange.py
@@ -23,10 +23,13 @@ LOG = logging.getLogger('feedhandler')
 class Exchange:
     id = NotImplemented
     symbol_endpoint = NotImplemented
+    websocket_endpoint = NotImplemented
+    sandbox_endpoint = NotImplemented
     _parse_symbol_data = NotImplemented
     websocket_channels = NotImplemented
     request_limit = NotImplemented
     valid_candle_intervals = NotImplemented
+    candle_interval_map = NotImplemented
     http_sync = HTTPSync()
 
     def __init__(self, config=None, sandbox=False, subaccount=None, **kwargs):