+++ cryptofeed/exchanges/fmfw.py
@@ -25,7 +25,9 @@ LOG = logging.getLogger('feedhandler')
 class FMFW(Feed):
     id = FMFW_id
     symbol_endpoint = 'https://api.fmfw.io/api/3/public/symbol'
+    websocket_endpoint = 'wss://api.fmfw.io/api/3/ws/public'
     valid_candle_intervals = {'1m', '3m', '5m', '15m', '30m', '1h', '4h', '1d', '1w', '1M'}
+    candle_interval_map = {'1m': 'M1', '3m': 'M3', '5m': 'M5', '15m': 'M15', '30m': 'M30', '1h': 'H1', '4h': 'H4', '1d': 'D1', '1w': 'D7', '1M': '1M'}
     websocket_channels = {
         L2_BOOK: 'orderbook/full',
         TRADES: 'trades',
@@ -49,12 +51,6 @@ class FMFW(Feed):
     @classmethod
     def timestamp_normalize(cls, ts: float) -> float:
         return ts / 1000.0
-
-    def __init__(self, **kwargs):
-        super().__init__('wss://api.fmfw.io/api/3/ws/public', **kwargs)
-        interval_map = {'1m': 'M1', '3m': 'M3', '5m': 'M5', '15m': 'M15', '30m': 'M30', '1h': 'H1', '4h': 'H4', '1d': 'D1', '1w': 'D7', '1M': '1M'}
-        self.candle_interval = interval_map[self.candle_interval]
-        self.normalize_interval = {value: key for key, value in interval_map.items()}
 
     def __reset(self):
         self._l2_book = {}
@@ -179,8 +175,8 @@ class FMFW(Feed):
                     self.id,
                     symbol,
                     u['t'] / 1000,
-                    u['t'] / 1000 + timedelta_str_to_sec(self.normalize_interval[interval]) - 0.1,
-                    self.normalize_interval[interval],
+                    u['t'] / 1000 + timedelta_str_to_sec(self.normalize_candle_interval[interval]) - 0.1,
+                    self.normalize_candle_interval[interval],
                     None,
                     Decimal(u['o']),
                     Decimal(u['c']),
@@ -214,6 +210,6 @@ class FMFW(Feed):
         for chan in self.subscription:
             await conn.write(json.dumps({"method": "subscribe",
                                          "params": {"symbols": self.subscription[chan]},
-                                         "ch": chan if chan != 'candles/' else chan + self.candle_interval,
+                                         "ch": chan if chan != 'candles/' else chan + self.candle_interval_map[self.candle_interval],
                                          "id": 1234
                                          }))