+++ cryptofeed/exchanges/kraken.py
@@ -28,8 +28,10 @@ LOG = logging.getLogger('feedhandler')
 class Kraken(Feed, KrakenRestMixin):
     id = KRAKEN
     valid_candle_intervals = {'1m', '5m', '15m', '30m', '1h', '4h', '1d', '1w', '15d'}
+    candle_interval_map = {'1m': 1, '5m': 5, '15m': 15, '30m': 30, '1h': 60, '4h': 240, '1d': 1440, '1w': 10080, '15d': 21600}
     valid_depths = [10, 25, 100, 500, 1000]
     symbol_endpoint = 'https://api.kraken.com/0/public/AssetPairs'
+    websocket_endpoint = 'wss://ws.kraken.com'
     websocket_channels = {
         L2_BOOK: 'book',
         TRADES: 'trade',
@@ -59,10 +61,7 @@ class Kraken(Feed, KrakenRestMixin):
         return ret, info
 
     def __init__(self, max_depth=1000, **kwargs):
-        super().__init__('wss://ws.kraken.com', max_depth=max_depth, **kwargs)
-        lookup = {'1m': 1, '5m': 5, '15m': 15, '30m': 30, '1h': 60, '4h': 240, '1d': 1440, '1w': 10080, '15d': 21600}
-        self.candle_interval = lookup[self.candle_interval]
-        self.normalize_interval = {value: key for key, value in lookup.items()}
+        super().__init__(max_depth=max_depth, **kwargs)
 
     def __reset(self):
         self._l2_book = {}
@@ -101,7 +100,7 @@ class Kraken(Feed, KrakenRestMixin):
 
             sub['depth'] = max_depth
         if self.exchange_channel_to_std(chan) == CANDLES:
-            sub['interval'] = self.candle_interval
+            sub['interval'] = self.candle_interval_map[self.candle_interval]
 
         await conn.write(json.dumps({
             "event": "subscribe",
@@ -200,7 +199,7 @@ class Kraken(Feed, KrakenRestMixin):
             pair,
             float(end) - (interval * 60),
             float(end),
-            self.normalize_interval[interval],
+            self.normalize_candle_interval[interval],
             count,
             Decimal(open),
             Decimal(close),