+++ tests/integration/rest/test_deribit.py
@@ -30,7 +30,7 @@ class TestDeribitRest:
     def test_trades(self):
         ret = []
         start = dt.now() - timedelta(days=5)
-        end = dt.now() - timedelta(days=4, hours=19)
+        end = dt.now() - timedelta(days=4, hours=18)
 
         for data in d.trades_sync('BTC-USD-PERP', start=start, end=end):
             ret.extend(data)