+++ CHANGES.md
@@ -7,6 +7,8 @@
  * Feature: Add support for sandbox/testnet on BinanceFutures.
  * Feature: New exchange - Crypto.com.
  * Bugfix: Fix MongoDB backend.
+ * Update: reduce code duplication for candle interval normalization.
+ * Update: Simplify code around address specification and selection when using sandbox/testnet.
 
 ### 2.1.0 (2021-11-14)
  * Bugfix: Update binance user data streams to use cdef types.